Create README.md
I created this project when i was doing a course on data analysis with Python. This project was created from all the concepts which i learnt from the course. Reliance Stock Market Data Analysis 2020

As we all our known with the word stock market in which stocks of the commodities, currencies & many more are traded/exchanged. I personally have a great interest in Stock market so i thought of working on it. I took this dataset from kaggle. I chose Reliance stocks for analysing.

This dataset consist of the reliance stocks highs and lows, opening price, previous closing price. In this dataset i have analyzed about their stocks by using pandas for cleaning and preparation and also i have used matplotlib and seaborn tools for visualizing of data.

Process - Data Cleaning and Preparation Data Exploration and Visualization